Attention casted to float in swin transformer

+++ SwinTransformer/models/swin_transformer_detectron.py
@@ -143,7 +143,7 @@ class WindowAttention(nn.Module):
 
         attn = self.attn_drop(attn)
 
-        x = (attn @ v).transpose(1, 2).reshape(B_, N, C)
+        x = (attn.float() @ v.float()).transpose(1, 2).reshape(B_, N, C)
         x = self.proj(x)
         x = self.proj_drop(x)
         return x
